requirements_test.txt
pytest==4.2.0
pytest-timeout==1.3.3
pydocstyle==3.0.0
pylint==2.2.2
pytest-aiohttp==0.3.0
pytest-cov==2.6.1
coveralls==1.2.0
pytest-sugar==0.9.2
flake8==3.7.5